sync cells passing to ReactGrid with store
